Remove second unneeded array
One std::array is enough to transform the iota initialized array to the
array containing the sum of proper factors of i, so this way we save 50%
of memory.

Also add -Wl,-rpath so that the linker will update the rpath of the
executable to know where to search for the TBB shared lib, in case it's
not in the LD_LIBRARY_PATH environment variable.